Move emitter to subfolder
