+++ lib/bloomFilter/emitter/BloomFilterEmitter.js
