+++ lib/bloomFilter/emitter/BloomFilterEmitterCollection.js
