+++ lib/grpcServer/handlers/getTransactionsByFilterHandlerFactory.js
@@ -3,7 +3,7 @@ const { MerkleBlock } = require('@dashevo/dashcore-lib');
 const { TransactionFilterResponse } = require('@dashevo/dapi-grpc');
 
 const TransactionHashesCache = require('../../transactionsFilter/TransactionHashesCache');
-const BloomFilterEmitter = require('../../bloomFilter/BloomFilterEmitter');
+const BloomFilterEmitter = require('../../bloomFilter/emitter/BloomFilterEmitter');
 
 /**
  *
